WIP on moving out renderers
